Ensure http and https are accounted for
The regex to clean the urls was not matching http prefixed urls.
They now are found in clean_the_url. We can subsequently use the
https version to search for a content id and if found, replace
the url with a valid version much in the same way as we are
removing chapters from url suffixes.