Prefix `EXECUTOR_NUMBER` to the test database name
We currently have Whitehall pinned to one agent in CI and this is
causing issues when two or more builds run concurrently. Each time a
build starts it drops the test dastabases and as these aren't currently
namespaced the already running build then fails.

This commit prefixes the DB names with the EXECUTOR_NUMBER environment
variable which should prevent this condition occurring.